updating utils after refractoring process
